requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.23
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.3
gunicorn==21.2.0
Flask-Session==0.5.0 